Use /q for VS 2010 by default instead of unattend template
VS 2010's unattend.ini files are OS-specific, so a given file will only work on the version of Windows
on which it was generated. In order to support installing VS 2010 on any OS out of the box, I have
changed this cookbook to use `/q` by default instead of `/unattend`. The `/q` option specifies a
silent installation with default options.

You may still specify an `unattend.ini` file if you wish to customize the installation.

This fixes issue #36.